services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["k-rea <kitahara@k-rea.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.1"
websockets = "^10.3"
SQLAlchemy = "^1.4.36"
GeoAlchemy2 = "^0.11.1"
psycopg2-binary = "^2.9.3"
SQLAlchemy-Utils = "^0.38.2"
Shapely = "^1.8.2"
strawberry-graphql = "^0.113.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
requests = "^2.27.1"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
alembic = "^1.7.7"
pytest-asyncio = "^0.18.3"

[tool.isort]
profile = "black"
line_length=119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
